[Refactor] AccountDeletionConfig belongs to Features namespace
