+++ config/initializers/features.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-AccountDeletionConfig.configure(Rails.configuration.three_scale.features.account_deletion)
+Features::AccountDeletionConfig.configure(Rails.configuration.three_scale.features.account_deletion)
