+++ test/workers/suspend_inactive_accounts_worker_test.rb
@@ -6,7 +6,7 @@ class SuspendInactiveAccountsWorkerTest < ActiveSupport::TestCase
   setup do
     account_inactivity = 50
     config = {'account_suspension' => 30, 'account_inactivity' => account_inactivity, 'contract_unpaid_time' => 70}
-    AccountDeletionConfig.configure(config)
+    Features::AccountDeletionConfig.configure(config)
 
     @accounts = {to_suspend: [], not_to_suspend: []}
 
@@ -58,7 +58,7 @@ class SuspendInactiveAccountsWorkerTest < ActiveSupport::TestCase
   end
 
   test 'it does not perform unless it has the valid configuration' do
-    AccountDeletionConfig.stubs(valid?: false)
+    Features::AccountDeletionConfig.stubs(valid?: false)
     SuspendInactiveAccountsWorker.new.perform
     (@accounts[:to_suspend] + @accounts[:not_to_suspend]).each { |account| refute account.reload.suspended? }
   end
